+++ spec/lib/pg_search/multisearch/rebuilder_spec.rb
@@ -3,7 +3,7 @@
 require "spec_helper"
 
 describe PgSearch::Multisearch::Rebuilder do
-  with_table "pg_search_documents", {}, &DOCUMENTS_SCHEMA
+  with_table "pg_search_documents", &DOCUMENTS_SCHEMA
 
   describe 'when initialized with a model that is not multisearchable' do
     with_model :not_multisearchable
@@ -221,6 +221,33 @@ describe PgSearch::Multisearch::Rebuilder do
             rebuilder.rebuild
 
             expect(record.pg_search_document).to be_present
+          end
+        end
+
+        context "when only additional_attributes is set" do
+          with_model :Model do
+            table do |t|
+              t.string :name
+            end
+
+            model do
+              include PgSearch::Model
+              multisearchable against: :name,
+                              additional_attributes: ->(obj) { { additional_attribute_column: "#{obj.class}::#{obj.id}" } }
+            end
+          end
+
+          it "calls update_pg_search_document on each record" do
+            record_1 = Model.create!(name: "record_1", id: 1)
+            record_2 = Model.create!(name: "record_2", id: 2)
+
+            PgSearch::Document.delete_all
+
+            rebuilder = PgSearch::Multisearch::Rebuilder.new(Model)
+            rebuilder.rebuild
+
+            expect(record_1.reload.pg_search_document.additional_attribute_column).to eq("Model::1")
+            expect(record_2.reload.pg_search_document.additional_attribute_column).to eq("Model::2")
           end
         end
       end