+++ spec/spec_helper.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
+require 'warning'
+# Ignore Ruby 2.7 warnings from Active Record
+Warning.ignore :keyword_separation
+
 require 'simplecov'
 SimpleCov.start
 
@@ -25,4 +29,7 @@ DOCUMENTS_SCHEMA = lambda do |t|
   t.belongs_to :searchable, polymorphic: true, index: true
   t.text :content
   t.timestamps null: false
+
+  # Used to test additional_attributes setup
+  t.text :additional_attribute_column
 end